+++ src/module_loader.py
@@ -1,5 +1,5 @@
 # Loads modules into the system
-
+import os
 import glob
 import warnings
 import importlib
@@ -16,9 +16,12 @@ class LoadModules(object):
         """
         self.modules = {}
         self.dmodules = {}
-        self.load_modules()
+        self.static_modules = {}
+        self.static_dmodules = {}
+        self.load_dynamic_modules()
+        self.load_static_modules()
 
-    def load_modules(self):
+    def load_dynamic_modules(self):
         """
         Loads modules into static class variables, 
         these can than be referenced easily.
@@ -27,13 +30,33 @@ class LoadModules(object):
         # loop and assign key and name
         warnings.filterwarnings('ignore', '.*Parent module*', )
         x = 1
-        for name in glob.glob('modules/*.py'):
+        path = os.path.join('src', 'dynamic_modules', '*.py')
+        for name in glob.glob(path):
             if name.endswith(".py") and ("__init__" not in name) \
                     and ("module_template" not in name):
                 module_name = name.replace("/", ".").rstrip('.py')
                 loaded_modules = self.dynamic_import(module_name)
                 self.modules[name] = loaded_modules
                 self.dmodules[x] = loaded_modules
+                x += 1
+
+    def load_static_modules(self):
+        """
+        Loads modules into static class variables,
+        these can than be referenced easily.
+        :return: NONE
+        """
+        # loop and assign key and name
+        warnings.filterwarnings('ignore', '.*Parent module*', )
+        x = 1
+        path = os.path.join('src', 'static_modules', '*.py')
+        for name in glob.glob(path):
+            if name.endswith(".py") and ("__init__" not in name) \
+                    and ("module_template" not in name):
+                module_name = name.replace("/", ".").rstrip('.py')
+                loaded_modules = self.dynamic_import(module_name)
+                self.static_modules[name] = loaded_modules
+                self.static_dmodules[x] = loaded_modules
                 x += 1
 
     def dynamic_import(self, module):
